Partially merge with Runtime.Bullet

+++ src/BitcoderCZ.Fancade.Runtime/EntryPoint.cs
@@ -8,17 +8,17 @@ namespace BitcoderCZ.Fancade.Runtime;
 internal readonly struct EntryPoint
 {
     public readonly int EnvironmentIndex;
-    public readonly ushort3 BlockPos;
+    public readonly int3 BlockPos;
     public readonly byte3 TerminalPos;
 
-    public EntryPoint(int environmentIndex, ushort3 blockPos, byte3 terminalPos)
+    public EntryPoint(int environmentIndex, int3 blockPos, byte3 terminalPos)
     {
         EnvironmentIndex = environmentIndex;
         BlockPos = blockPos;
         TerminalPos = terminalPos;
     }
 
-    public void Deconstruct(out int environmentIndex, out ushort3 blockPos, out byte3 terminalPos)
+    public void Deconstruct(out int environmentIndex, out int3 blockPos, out byte3 terminalPos)
     {
         environmentIndex = EnvironmentIndex;
         blockPos = BlockPos;
